style: use size_t over unsigned for count/capacity measurements

+++ source/PlayerInfo.cpp
@@ -763,7 +763,7 @@ const vector<shared_ptr<Ship>> &PlayerInfo::Ships() const
 std::map<const shared_ptr<Ship>, const string> PlayerInfo::FlightCheck() const
 {
 	// Count of all bay types in the active fleet.
-	auto bayCount = map<string, unsigned>{
+	auto bayCount = map<string, size_t>{
 		{"Fighter", 0},
 		{"Drone", 0},
 	};
